backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.3.1
pandas==2.3.3
reportlab==4.4.5
python-dotenv==1.2.1
